feat: add nested error boundary
